# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PARTYMAX",
        "description": "Official API for 'PARTYMAX' App",
        "contact": {
            "email": "admin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.partymax/api",
            "description": "API"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Lists available Roles",
                "description": "Gets all available Roles resources",
                "operationId": "6e28bdec08d7292bde33180ca5928718",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search Text",
                        "schema": {
                            "description": "Search text to find users",
                            "type": "string",
                            "format": "text"
                        }
                    },
                    {
                        "name": "orderByField",
                        "in": "query",
                        "description": "Order by field the query",
                        "schema": {
                            "description": "Order by field the query",
                            "type": "string",
                            "format": "text"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Default query order 'asc'",
                        "schema": {
                            "description": "Default query order asc",
                            "type": "string",
                            "format": "text"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of rows per page",
                        "schema": {
                            "description": "number of rows per page",
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "schema": {
                            "description": "number of page",
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a new Role",
                "description": "Create a new Role",
                "operationId": "b47f38c8385567ad7a473d82fb6ebf63",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "permissions"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name Role.",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "permissions": {
                                        "description": "Assigned permissions array.",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Some was wrong",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update a Role",
                "description": "Update a Role",
                "operationId": "e8c67c50f2f7ea7a8a12c812fc3ef2a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The role id",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a role",
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "permissions"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name role.",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "permissions": {
                                        "description": "Assigned permissions array.",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Some was wrong",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete a Role",
                "description": "Delete a Role",
                "operationId": "f5117dc2943b621bdb53820fcc72b457",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The role id",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a role",
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/roles/role/all": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Lists available Roles Array",
                "description": "Gets all available Roles Array resources",
                "operationId": "d29356d9fd8703d11a136e08c818ea5f",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}